Fix compilation on platforms with i32 c_long

+++ vm/src/stdlib/os.rs
@@ -1932,7 +1932,7 @@ mod _os {
                 Err(io::Error::from(Errno::last()).into_pyexception(vm))
             }
         } else {
-            Ok(Some(raw))
+            Ok(Some(raw as i64))
         }
     }
